requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9
pydantic>=2.0
pydantic-settings>=2.0
sse-starlette>=2.0.0
litellm>=1.40.0
qdrant-client>=1.9.0
openai>=1.0.0
tavily-python>=0.4.0
httpx>=0.27.0
pyyaml>=6.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
pymupdf>=1.25.0
rank-bm25>=0.2.2
bcrypt>=4.0.0
PyJWT>=2.8.0
